+++ .travis.yml
@@ -1,0 +1,14 @@
+language: java
+
+jdk:
+  - openjdk8
+
+script:
+  - ./gradlew integration
+
+# in case swagger specs generated by integration tests:
+before_install:
+  - curl https://raw.githubusercontent.com/hmcts/reform-api-docs/master/bin/publish-swagger-docs-dockerless.sh > publish-swagger-docs.sh
+
+after_success:
+  - test "$TRAVIS_BRANCH" = "master" && test "$TRAVIS_PULL_REQUEST" = "false" && sh ./publish-swagger-docs.sh
