+++ Jenkinsfile_CNP
@@ -1,0 +1,58 @@
+#!groovy
+properties([
+    [$class: 'GithubProjectProperty', projectUrlStr: 'https://github.com/hmcts/timed-event-service.git'],
+    pipelineTriggers([[$class: 'GitHubPushTrigger']])
+])
+
+
+@Library("Infrastructure")
+
+def type = "java"
+def product = "ia"
+def component = "timed-event-service"
+
+def secrets = [
+    'ia-${env}': [
+
+        secret('idam-client-id', 'IA_IDAM_CLIENT_ID'),
+        secret('idam-secret', 'IA_IDAM_SECRET'),
+        secret('idam-redirect-uri', 'IA_IDAM_REDIRECT_URI'),
+        secret('s2s-secret', 'IA_S2S_SECRET'),
+        secret('s2s-microservice', 'IA_S2S_MICROSERVICE'),
+        secret('ccd-url', 'CCD_URL'),
+        secret('ccd-gw-url', 'CCD_GW_URL'),
+        secret('dm-url', 'DM_URL'),
+        secret('idam-url', 'IDAM_URL'),
+        secret('s2s-url', 'S2S_URL')
+
+    ]
+]
+
+static LinkedHashMap<String, Object> secret(String secretName, String envVar) {
+    [$class     : 'AzureKeyVaultSecret',
+     secretType : 'Secret',
+     name       : secretName,
+     version    : '',
+     envVariable: envVar
+    ]
+}
+
+withPipeline(type, product, component) {
+    installCharts()
+
+    if ("demo" != "${env.BRANCH_NAME}") {
+        enableAksStagingDeployment()
+        disableLegacyDeployment()
+    }
+    loadVaultSecrets(secrets)
+
+    after('functionalTest:preview') {
+        steps.archiveArtifacts allowEmptyArchive: true, artifacts: '**/site/serenity/**/*'
+    }
+
+    after('functionalTest:aat') {
+        steps.archiveArtifacts allowEmptyArchive: true, artifacts: '**/site/serenity/**/*'
+    }
+
+    enableSlackNotifications('#ia-tech')
+}
