+++ Jenkinsfile_nightly
@@ -1,0 +1,51 @@
+#!groovy
+
+properties([
+    // H allow predefined but random minute see https://en.wikipedia.org/wiki/Cron#Non-standard_characters
+    pipelineTriggers([cron('H 22 * * *')]),
+    parameters([
+        string(name: 'URL_TO_TEST', defaultValue: 'http://ia-timed-event-service-aat.service.core-compute-aat.internal', description: 'The URL you want to run these tests against'),
+    ])
+])
+
+@Library("Infrastructure")
+
+def type = "java"
+def product = "ia"
+def component = "timed-event-service"
+
+def secrets = [
+    'ia-${env}': [
+
+        secret('idam-client-id', 'IA_IDAM_CLIENT_ID'),
+        secret('idam-secret', 'IA_IDAM_SECRET'),
+        secret('idam-redirect-uri', 'IA_IDAM_REDIRECT_URI'),
+        secret('s2s-secret', 'IA_S2S_SECRET'),
+        secret('s2s-microservice', 'IA_S2S_MICROSERVICE'),
+        secret('ccd-url', 'CCD_URL'),
+        secret('ccd-gw-url', 'CCD_GW_URL'),
+        secret('dm-url', 'DM_URL'),
+        secret('idam-url', 'IDAM_URL'),
+        secret('s2s-url', 'S2S_URL')
+
+    ]
+]
+
+static LinkedHashMap<String, Object> secret(String secretName, String envVar) {
+    [$class     : 'AzureKeyVaultSecret',
+     secretType : 'Secret',
+     name       : secretName,
+     version    : '',
+     envVariable: envVar
+    ]
+}
+
+withNightlyPipeline(type, product, component) {
+    env.TEST_URL = params.URL_TO_TEST
+
+    loadVaultSecrets(secrets)
+
+    enableSecurityScan()
+    enableMutationTest()
+    enableSlackNotifications('#ia-tech')
+}
