+++ charts/ia-timed-event-service/Chart.yaml
@@ -1,0 +1,8 @@
+apiVersion: v1
+name: ia-timed-event-service
+home: https://github.com/hmcts/timed-event-service
+version: 0.0.1
+description: Timed Event Service
+maintainers:
+  - name: HMCTS Immigration & Asylum Team
+    email: ImmigrationandAsylum@HMCTS.NET
