+++ charts/ia-timed-event-service/requirements.yaml
@@ -1,0 +1,4 @@
+dependencies:
+  - name: java
+    version: ~2.9.1
+    repository: '@hmctspublic'
