+++ charts/ia-timed-event-service/values.aat.template.yaml
@@ -1,0 +1,4 @@
+java:
+  # Don't modify below here
+  image: ${IMAGE_NAME}
+  ingressHost: ${SERVICE_FQDN}
