+++ charts/ia-timed-event-service/values.yaml
@@ -1,0 +1,19 @@
+java:
+  image: 'https://hmctspublic.azurecr.io/ia/timed-event-service:latest'
+  applicationPort: 8095
+  aadIdentityName: ia
+  environment:
+    IDAM_URL: "http://idam-api-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal"
+    S2S_URL: "http://rpe-service-auth-provider-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal"
+    CCD_URL: "http://ccd-data-store-api-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal"
+  keyVaults:
+    ia:
+      secrets:
+        - system-username
+        - system-password
+        - idam-client-id
+        - idam-secret
+        - idam-redirect-uri
+        - s2s-secret
+        - s2s-microservice
+        - AppInsightsInstrumentationKey
