--- src/main/java/uk/gov/hmcts/reform/demo/config/SwaggerConfiguration.java
@@ -1,26 +1,0 @@
-package uk.gov.hmcts.reform.demo.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-import uk.gov.hmcts.reform.demo.Application;
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfiguration {
-
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-            .useDefaultResponseMessages(false)
-            .select()
-            .apis(RequestHandlerSelectors.basePackage(Application.class.getPackage().getName() + ".controllers"))
-            .paths(PathSelectors.any())
-            .build();
-    }
-
-}
