+++ src/main/java/uk/gov/hmcts/reform/timed/event/controllers/RootController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.controllers;
+package uk.gov.hmcts.reform.timed.event.controllers;
 
 import static org.springframework.http.ResponseEntity.ok;
 
@@ -6,9 +6,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-/**
- * Default endpoints per application.
- */
 @RestController
 public class RootController {
 
@@ -23,6 +20,6 @@ public class RootController {
      */
     @GetMapping("/")
     public ResponseEntity<String> welcome() {
-        return ok("Welcome to spring-boot-template");
+        return ok("Welcome to Timed Event Service");
     }
 }
